feat(spaces): upsert
Reverts bcd7caf
