+++ packages/frontend/src/components/common/dialogs/BaseDialog.vue
@@ -10,7 +10,7 @@ const { setIsInputFocused } = useCommands();
 const dialogComponents = {
   [DIALOGS.CONFIRM]: () => import('./ConfirmDialog.vue'),
   [DIALOGS.CREATE_CARD]: () => import('./CreateCardDialog.vue'),
-  [DIALOGS.CREATE_SPACE]: () => import('./CreateSpaceDialog.vue'),
+  [DIALOGS.UPSERT_SPACE]: () => import('./UpsertSpaceDialog.vue'),
   [DIALOGS.UPSERT_LIST]: () => import('./UpsertListDialog.vue'),
   [DIALOGS.UPSERT_LIST_STAGE]: () => import('./UpsertListStageDialog.vue'),
   [DIALOGS.REMOVE_LIST_STAGE]: () => import('./RemoveListStageDialog.vue'),
