+++ packages/frontend/src/components/common/dialogs/UpsertSpaceDialog.vue
@@ -1,0 +1,119 @@
+<script setup lang="ts">
+import { useSpacesService } from '@/composables/services/useSpacesService';
+import { useWorkspaceStore } from '@/stores/workspace';
+import { type VForm } from 'vuetify/components';
+import validationUtils from '@/utils/validation';
+import type { Space } from '@/components/project-management/spaces/types';
+import { useSnackbarStore } from '@/stores/snackbar';
+import { useDialogStore } from '@/stores/dialog';
+import { DIALOGS } from './types';
+import { cloneDeep } from 'lodash';
+
+const workspaceStore = useWorkspaceStore();
+const { selectedWorkspace } = storeToRefs(workspaceStore);
+const spacesService = useSpacesService();
+const dialog = useDialogStore();
+const { rules } = validationUtils;
+const { showSnackbar } = useSnackbarStore();
+
+const currentDialogIndex = computed(() =>
+  dialog.getDialogIndex(DIALOGS.UPSERT_SPACE)
+);
+const currentDialog = computed(() => dialog.dialogs[currentDialogIndex.value]);
+const space = computed<Space>(() => currentDialog.value.data.space);
+
+const spaceForm = ref<VForm>();
+const spaceDto = ref<Partial<Space>>({
+  name: space.value?.name,
+  // TODO: Allow to update workspace?
+  workspaceId: selectedWorkspace.value?.id,
+});
+
+const { mutateAsync: createSpace, isPending: isCreating } =
+  spacesService.useCreateSpaceMutation();
+const { mutateAsync: updateSpace, isPending: isUpdating } =
+  spacesService.useUpdateSpaceMutation();
+
+async function handleSubmitForm() {
+  const isValid = await spaceForm.value?.validate();
+  if (!isValid?.valid) return;
+
+  try {
+    switch (currentDialog.value.data.mode) {
+      case 'Create':
+        createSpace(spaceDto.value).then(() => {
+          dialog.closeDialog(currentDialogIndex.value);
+        });
+        break;
+
+      case 'Update':
+        // - nothing has changed
+        // NOTE: Thanks to validation, we already know we aren't getting an empty string.
+        if (spaceDto.value.name === space.value.name) {
+          dialog.closeDialog(currentDialogIndex.value);
+          break;
+        }
+
+        {
+          const spaceCopy = cloneDeep(space.value);
+          spaceCopy.name = spaceDto.value.name!;
+          updateSpace(spaceCopy).then(() => {
+            dialog.closeDialog(currentDialogIndex.value);
+          });
+        }
+        break;
+    }
+  } catch {
+    showSnackbar({
+      message: 'Something went wrong, please try again.',
+      color: 'error',
+      timeout: 5000,
+    });
+  }
+}
+</script>
+
+<template>
+  <v-card color="surface" elevation="24" :loading="isCreating || isUpdating">
+    <div class="d-flex align-center ps-0 pa-4">
+      <v-card-subtitle>
+        {{ currentDialog?.data.mode }}
+        Space
+        {{ !currentDialog?.data.space ? 'in ' + selectedWorkspace?.name : '' }}
+      </v-card-subtitle>
+      <v-spacer />
+      <base-icon-btn
+        icon="mdi-close"
+        color="default"
+        @click="dialog.closeDialog(currentDialogIndex)"
+      />
+    </div>
+    <v-form
+      ref="spaceForm"
+      @submit.prevent="handleSubmitForm"
+      validate-on="submit"
+    >
+      <div class="pa-4 py-0">
+        <v-text-field
+          v-model="spaceDto.name"
+          :rules="[rules.required]"
+          label="Name*"
+          autofocus
+        />
+      </div>
+      <v-card-actions class="d-flex justify-start align-center py-0 px-4">
+        <v-spacer />
+
+        <v-btn
+          density="comfortable"
+          variant="flat"
+          class="text-caption px-4 ms-4"
+          type="submit"
+          :loading="isCreating || isUpdating"
+        >
+          {{ currentDialog?.data.mode }}
+        </v-btn>
+      </v-card-actions>
+    </v-form>
+  </v-card>
+</template>
