+++ packages/frontend/src/components/common/dialogs/types.ts
@@ -3,7 +3,7 @@ export enum DIALOGS {
   CONFIRM = 'ConfirmDialog',
   ONBOARDING = 'OnboardingDialog',
   SETTINGS = 'SettingsDialog',
-  CREATE_SPACE = 'CreateSpaceDialog',
+  UPSERT_SPACE = 'UpsertSpaceDialog',
   UPSERT_LIST = 'UpsertListDialog',
   CREATE_WORKSPACE = 'CreateWorkspaceDialog',
   CREATE_VIEW = 'CreateViewDialog',
@@ -16,7 +16,7 @@ export enum DIALOGS {
 
 export const DIALOG_WIDTHS = {
   [DIALOGS.CREATE_CARD]: 700,
-  [DIALOGS.CREATE_SPACE]: 400,
+  [DIALOGS.UPSERT_SPACE]: 400,
   [DIALOGS.UPSERT_LIST]: 400,
   [DIALOGS.CREATE_VIEW]: 400,
   [DIALOGS.CONFIRM]: 500,
